+++ paq8pxd.cpp
@@ -24,7 +24,7 @@
 #define SIMD_GET_SSE  //uncomment to use SSE2 in ContexMap
 //#define MT            //uncomment for multithreading, compression only. Handled by CMake and gcc when -DMT is passed.
 #ifndef DISABLE_SM
-#define SM              // For faster statemap
+//#define SM              // For faster statemap
 #endif
 //#define VERBOSE         // Show extra info
 
@@ -114,6 +114,10 @@ inline int max(int a, int b) {return a<b?b:a;}
 #include "prt/blockdata.hpp"
 #include "prt/logistic.hpp"
 #include "prt/mixer.hpp"
+#include "prt/APM1.hpp"
+#include "prt/statemap.hpp"
+#include "prt/statetable.hpp"
+#include "prt/APM.hpp"
 
 /////////////////////// Global context /////////////////////////
 U8 level=DEFAULT_OPTION;  // Compression level 0 to 15
@@ -230,417 +234,10 @@ const U8 WRT_wrd1[256]={
  192,192,192,192,192,192,192,192, 192,192,192,192,192,192,192,192,
  };
 const U32 tri[4]={0,4,3,7}, trj[4]={0,6,6,12};
+
+int dt[1024];  // i -> 16K/(i+i+3)
 #define PNGFlag (1<<31)
 #define GrayFlag (1<<30)
-
-///////////////////////// state table ////////////////////////
-
-// State table:
-//   nex(state, 0) = next state if bit y is 0, 0 <= state < 256
-//   nex(state, 1) = next state if bit y is 1
-//   nex(state, 2) = number of zeros in bit history represented by state
-//   nex(state, 3) = number of ones represented
-//
-// States represent a bit history within some context.
-// State 0 is the starting state (no bits seen).
-// States 1-30 represent all possible sequences of 1-4 bits.
-// States 31-252 represent a pair of counts, (n0,n1), the number
-//   of 0 and 1 bits respectively.  If n0+n1 < 16 then there are
-//   two states for each pair, depending on if a 0 or 1 was the last
-//   bit seen.
-// If n0 and n1 are too large, then there is no state to represent this
-// pair, so another state with about the same ratio of n0/n1 is substituted.
-// Also, when a bit is observed and the count of the opposite bit is large,
-// then part of this count is discarded to favor newer data over old.
-
-#if 1 // change to #if 0 to generate this table at run time (4% slower)
-static const U8 State_table[256][4]={
-  {  1,  2, 0, 0},{  3,  5, 1, 0},{  4,  6, 0, 1},{  7, 10, 2, 0}, // 0-3
-  {  8, 12, 1, 1},{  9, 13, 1, 1},{ 11, 14, 0, 2},{ 15, 19, 3, 0}, // 4-7
-  { 16, 23, 2, 1},{ 17, 24, 2, 1},{ 18, 25, 2, 1},{ 20, 27, 1, 2}, // 8-11
-  { 21, 28, 1, 2},{ 22, 29, 1, 2},{ 26, 30, 0, 3},{ 31, 33, 4, 0}, // 12-15
-  { 32, 35, 3, 1},{ 32, 35, 3, 1},{ 32, 35, 3, 1},{ 32, 35, 3, 1}, // 16-19
-  { 34, 37, 2, 2},{ 34, 37, 2, 2},{ 34, 37, 2, 2},{ 34, 37, 2, 2}, // 20-23
-  { 34, 37, 2, 2},{ 34, 37, 2, 2},{ 36, 39, 1, 3},{ 36, 39, 1, 3}, // 24-27
-  { 36, 39, 1, 3},{ 36, 39, 1, 3},{ 38, 40, 0, 4},{ 41, 43, 5, 0}, // 28-31
-  { 42, 45, 4, 1},{ 42, 45, 4, 1},{ 44, 47, 3, 2},{ 44, 47, 3, 2}, // 32-35
-  { 46, 49, 2, 3},{ 46, 49, 2, 3},{ 48, 51, 1, 4},{ 48, 51, 1, 4}, // 36-39
-  { 50, 52, 0, 5},{ 53, 43, 6, 0},{ 54, 57, 5, 1},{ 54, 57, 5, 1}, // 40-43
-  { 56, 59, 4, 2},{ 56, 59, 4, 2},{ 58, 61, 3, 3},{ 58, 61, 3, 3}, // 44-47
-  { 60, 63, 2, 4},{ 60, 63, 2, 4},{ 62, 65, 1, 5},{ 62, 65, 1, 5}, // 48-51
-  { 50, 66, 0, 6},{ 67, 55, 7, 0},{ 68, 57, 6, 1},{ 68, 57, 6, 1}, // 52-55
-  { 70, 73, 5, 2},{ 70, 73, 5, 2},{ 72, 75, 4, 3},{ 72, 75, 4, 3}, // 56-59
-  { 74, 77, 3, 4},{ 74, 77, 3, 4},{ 76, 79, 2, 5},{ 76, 79, 2, 5}, // 60-63
-  { 62, 81, 1, 6},{ 62, 81, 1, 6},{ 64, 82, 0, 7},{ 83, 69, 8, 0}, // 64-67
-  { 84, 71, 7, 1},{ 84, 71, 7, 1},{ 86, 73, 6, 2},{ 86, 73, 6, 2}, // 68-71
-  { 44, 59, 5, 3},{ 44, 59, 5, 3},{ 58, 61, 4, 4},{ 58, 61, 4, 4}, // 72-75
-  { 60, 49, 3, 5},{ 60, 49, 3, 5},{ 76, 89, 2, 6},{ 76, 89, 2, 6}, // 76-79
-  { 78, 91, 1, 7},{ 78, 91, 1, 7},{ 80, 92, 0, 8},{ 93, 69, 9, 0}, // 80-83
-  { 94, 87, 8, 1},{ 94, 87, 8, 1},{ 96, 45, 7, 2},{ 96, 45, 7, 2}, // 84-87
-  { 48, 99, 2, 7},{ 48, 99, 2, 7},{ 88,101, 1, 8},{ 88,101, 1, 8}, // 88-91
-  { 80,102, 0, 9},{103, 69,10, 0},{104, 87, 9, 1},{104, 87, 9, 1}, // 92-95
-  {106, 57, 8, 2},{106, 57, 8, 2},{ 62,109, 2, 8},{ 62,109, 2, 8}, // 96-99
-  { 88,111, 1, 9},{ 88,111, 1, 9},{ 80,112, 0,10},{113, 85,11, 0}, // 100-103
-  {114, 87,10, 1},{114, 87,10, 1},{116, 57, 9, 2},{116, 57, 9, 2}, // 104-107
-  { 62,119, 2, 9},{ 62,119, 2, 9},{ 88,121, 1,10},{ 88,121, 1,10}, // 108-111
-  { 90,122, 0,11},{123, 85,12, 0},{124, 97,11, 1},{124, 97,11, 1}, // 112-115
-  {126, 57,10, 2},{126, 57,10, 2},{ 62,129, 2,10},{ 62,129, 2,10}, // 116-119
-  { 98,131, 1,11},{ 98,131, 1,11},{ 90,132, 0,12},{133, 85,13, 0}, // 120-123
-  {134, 97,12, 1},{134, 97,12, 1},{136, 57,11, 2},{136, 57,11, 2}, // 124-127
-  { 62,139, 2,11},{ 62,139, 2,11},{ 98,141, 1,12},{ 98,141, 1,12}, // 128-131
-  { 90,142, 0,13},{143, 95,14, 0},{144, 97,13, 1},{144, 97,13, 1}, // 132-135
-  { 68, 57,12, 2},{ 68, 57,12, 2},{ 62, 81, 2,12},{ 62, 81, 2,12}, // 136-139
-  { 98,147, 1,13},{ 98,147, 1,13},{100,148, 0,14},{149, 95,15, 0}, // 140-143
-  {150,107,14, 1},{150,107,14, 1},{108,151, 1,14},{108,151, 1,14}, // 144-147
-  {100,152, 0,15},{153, 95,16, 0},{154,107,15, 1},{108,155, 1,15}, // 148-151
-  {100,156, 0,16},{157, 95,17, 0},{158,107,16, 1},{108,159, 1,16}, // 152-155
-  {100,160, 0,17},{161,105,18, 0},{162,107,17, 1},{108,163, 1,17}, // 156-159
-  {110,164, 0,18},{165,105,19, 0},{166,117,18, 1},{118,167, 1,18}, // 160-163
-  {110,168, 0,19},{169,105,20, 0},{170,117,19, 1},{118,171, 1,19}, // 164-167
-  {110,172, 0,20},{173,105,21, 0},{174,117,20, 1},{118,175, 1,20}, // 168-171
-  {110,176, 0,21},{177,105,22, 0},{178,117,21, 1},{118,179, 1,21}, // 172-175
-  {110,180, 0,22},{181,115,23, 0},{182,117,22, 1},{118,183, 1,22}, // 176-179
-  {120,184, 0,23},{185,115,24, 0},{186,127,23, 1},{128,187, 1,23}, // 180-183
-  {120,188, 0,24},{189,115,25, 0},{190,127,24, 1},{128,191, 1,24}, // 184-187
-  {120,192, 0,25},{193,115,26, 0},{194,127,25, 1},{128,195, 1,25}, // 188-191
-  {120,196, 0,26},{197,115,27, 0},{198,127,26, 1},{128,199, 1,26}, // 192-195
-  {120,200, 0,27},{201,115,28, 0},{202,127,27, 1},{128,203, 1,27}, // 196-199
-  {120,204, 0,28},{205,115,29, 0},{206,127,28, 1},{128,207, 1,28}, // 200-203
-  {120,208, 0,29},{209,125,30, 0},{210,127,29, 1},{128,211, 1,29}, // 204-207
-  {130,212, 0,30},{213,125,31, 0},{214,137,30, 1},{138,215, 1,30}, // 208-211
-  {130,216, 0,31},{217,125,32, 0},{218,137,31, 1},{138,219, 1,31}, // 212-215
-  {130,220, 0,32},{221,125,33, 0},{222,137,32, 1},{138,223, 1,32}, // 216-219
-  {130,224, 0,33},{225,125,34, 0},{226,137,33, 1},{138,227, 1,33}, // 220-223
-  {130,228, 0,34},{229,125,35, 0},{230,137,34, 1},{138,231, 1,34}, // 224-227
-  {130,232, 0,35},{233,125,36, 0},{234,137,35, 1},{138,235, 1,35}, // 228-231
-  {130,236, 0,36},{237,125,37, 0},{238,137,36, 1},{138,239, 1,36}, // 232-235
-  {130,240, 0,37},{241,125,38, 0},{242,137,37, 1},{138,243, 1,37}, // 236-239
-  {130,244, 0,38},{245,135,39, 0},{246,137,38, 1},{138,247, 1,38}, // 240-243
-  {140,248, 0,39},{249,135,40, 0},{250, 69,39, 1},{ 80,251, 1,39}, // 244-247
-  {140,252, 0,40},{249,135,41, 0},{250, 69,40, 1},{ 80,251, 1,40}, // 248-251
-  {140,252, 0,41}};  // 252, 253-255 are reserved
-
-#define nex(state,sel) State_table[state][sel]
-
-// The code used to generate the above table at run time (4% slower).
-// To print the table, uncomment the 4 lines of print statements below.
-// In this code x,y = n0,n1 is the number of 0,1 bits represented by a state.
-#else
-
-class StateTable {
-  Array<U8> ns;  // state*4 -> next state if 0, if 1, n0, n1
-  enum {B=5, N=64}; // sizes of b, t
-  static const int b[B];  // x -> max y, y -> max x
-  static U8 t[N][N][2];  // x,y -> state number, number of states
-  int num_states(int x, int y);  // compute t[x][y][1]
-  void discount(int& x);  // set new value of x after 1 or y after 0
-  void next_state(int& x, int& y, int b);  // new (x,y) after bit b
-public:
-  int operator()(int state, int sel) {return ns[state*4+sel];}
-  StateTable();
-} nex;
-
-const int StateTable::b[B]={42,41,13,6,5};  // x -> max y, y -> max x
-U8 StateTable::t[N][N][2];
-
-int StateTable::num_states(int x, int y) {
-  if (x<y) return num_states(y, x);
-  if (x<0 || y<0 || x>=N || y>=N || y>=B || x>=b[y]) return 0;
-
-  // States 0-30 are a history of the last 0-4 bits
-  if (x+y<=4) {  // x+y choose x = (x+y)!/x!y!
-    int r=1;
-    for (int i=x+1; i<=x+y; ++i) r*=i;
-    for (int i=2; i<=y; ++i) r/=i;
-    return r;
-  }
-
-  // States 31-255 represent a 0,1 count and possibly the last bit
-  // if the state is reachable by either a 0 or 1.
-  else
-    return 1+(y>0 && x+y<16);
-}
-
-// New value of count x if the opposite bit is observed
-void StateTable::discount(int& x) {
-  if (x>2) x=ilog(x)/6-1;
-}
-
-// compute next x,y (0 to N) given input b (0 or 1)
-void StateTable::next_state(int& x, int& y, int b) {
-  if (x<y)
-    next_state(y, x, 1-b);
-  else {
-    if (b) {
-      ++y;
-      discount(x);
-    }
-    else {
-      ++x;
-      discount(y);
-    }
-    while (!t[x][y][1]) {
-      if (y<2) --x;
-      else {
-        x=(x*(y-1)+(y/2))/y;
-        --y;
-      }
-    }
-  }
-}
-
-// Initialize next state table ns[state*4] -> next if 0, next if 1, x, y
-StateTable::StateTable(): ns(1024) {
-
-  // Assign states
-  int state=0;
-  for (int i=0; i<256; ++i) {
-    for (int y=0; y<=i; ++y) {
-      int x=i-y;
-      int n=num_states(x, y);
-      if (n) {
-        t[x][y][0]=state;
-        t[x][y][1]=n;
-        state+=n;
-      }
-    }
-  }
-
-  // Print/generate next state table
-  state=0;
-  for (int i=0; i<N; ++i) {
-    for (int y=0; y<=i; ++y) {
-      int x=i-y;
-      for (int k=0; k<t[x][y][1]; ++k) {
-        int x0=x, y0=y, x1=x, y1=y;  // next x,y for input 0,1
-        int ns0=0, ns1=0;
-        if (state<15) {
-          ++x0;
-          ++y1;
-          ns0=t[x0][y0][0]+state-t[x][y][0];
-          ns1=t[x1][y1][0]+state-t[x][y][0];
-          if (x>0) ns1+=t[x-1][y+1][1];
-          ns[state*4]=ns0;
-          ns[state*4+1]=ns1;
-          ns[state*4+2]=x;
-          ns[state*4+3]=y;
-        }
-        else if (t[x][y][1]) {
-          next_state(x0, y0, 0);
-          next_state(x1, y1, 1);
-          ns[state*4]=ns0=t[x0][y0][0];
-          ns[state*4+1]=ns1=t[x1][y1][0]+(t[x1][y1][1]>1);
-          ns[state*4+2]=x;
-          ns[state*4+3]=y;
-        }
-          // uncomment to print table above
-//        printf("{%3d,%3d,%2d,%2d},", ns[state*4], ns[state*4+1],
-//          ns[state*4+2], ns[state*4+3]);
-//        if (state%4==3) printf(" // %d-%d\n  ", state-3, state);
-        assert(state>=0 && state<256);
-        assert(t[x][y][1]>0);
-        assert(t[x][y][0]<=state);
-        assert(t[x][y][0]+t[x][y][1]>state);
-        assert(t[x][y][1]<=6);
-        assert(t[x0][y0][1]>0);
-        assert(t[x1][y1][1]>0);
-        assert(ns0-t[x0][y0][0]<t[x0][y0][1]);
-        assert(ns0-t[x0][y0][0]>=0);
-        assert(ns1-t[x1][y1][0]<t[x1][y1][1]);
-        assert(ns1-t[x1][y1][0]>=0);
-        ++state;
-      }
-    }
-  }
-//  printf("%d states\n", state); exit(0);  // uncomment to print table above
-}
-
-#endif
-
-
-
-
-
-//////////////////////////// APM1 //////////////////////////////
-
-// APM1 maps a probability and a context into a new probability
-// that bit y will next be 1.  After each guess it updates
-// its state to improve future guesses.  Methods:
-//
-// APM1 a(N) creates with N contexts, uses 66*N bytes memory.
-// a.p(pr, cx, rate=7) returned adjusted probability in context cx (0 to
-//   N-1).  rate determines the learning rate (smaller = faster, default 7).
-//   Probabilities are scaled 12 bits (0-4095).
-
-class APM1 {
-  int index;     // last p, context
-  const int N;   // number of contexts
-  Array<U16> t;  // [N][33]:  p, context -> p
-  BlockData& x;
-public:
-  APM1(int n,BlockData& x);
-  int p(int pr=2048, int cxt=0, int rate=7) {
-    assert(pr>=0 && pr<4096 && cxt>=0 && cxt<N && rate>0 && rate<32);
-    pr=stretch(pr);
-    int g=(x.y<<16)+(x.y<<rate)-x.y-x.y;
-    t[index] += (g-t[index]) >> rate;
-    t[index+1] += (g-t[index+1]) >> rate;
-    const int w=pr&127;  // interpolation weight (33 points)
-    index=((pr+2048)>>7)+cxt*33;
-    return (t[index]*(128-w)+t[index+1]*w) >> 11;
-  }
-};
-
-// maps p, cxt -> p initially
-APM1::APM1(int n,BlockData& bd): index(0), N(n), t(n*33),x(bd) {
-  for (int i=0; i<N; ++i)
-    for (int j=0; j<33; ++j)
-      t[i*33+j] = i==0 ? squash((j-16)*128)*16 : t[j];
-}
-
-//////////////////////////// StateMap, APM //////////////////////////
-
-// A StateMap maps a context to a probability.  Methods:
-//
-// Statemap sm(n) creates a StateMap with n contexts using 4*n bytes memory.
-// sm.p(y, cx, limit) converts state cx (0..n-1) to a probability (0..4095).
-//     that the next y=1, updating the previous prediction with y (0..1).
-//     limit (1..1023, default 1023) is the maximum count for computing a
-//     prediction.  Larger values are better for stationary sources.
-
-static int dt[1024];  // i -> 16K/(i+i+3)
-class StateMap {
-protected:
-  const int N;  // Number of contexts
-  int cxt;      // Context of last prediction
-  Array<U32> t;       // cxt -> prediction in high 22 bits, count in low 10 bits
-  inline void update(const int y, int limit) {
-    assert(cxt>=0 && cxt<N);
-    assert(y==0 || y==1);
-    U32 *p=&t[cxt], p0=p[0];
-#ifdef SM   
-    int n=p0&1023, pr=p0>>13;  // count, prediction
-    p0+=(n<limit);
-    U32 err=(y<<19)-pr;
-    p0+=((err))*dt[n]&0xfffffc00;
-
-#else
-    int n=p0&1023, pr=p0>>10;  // count, prediction
-    p0+=(n<limit);
-    p0+=(((y<<22)-pr)>>3)*dt[n]&0xfffffc00;
-#endif
-    p[0]=p0;
-  }
-  
-inline void update1(const int y, int limit) {
-    assert(cxt>=0 && cxt<N);
-    assert(y==0 || y==1);
-    U32   p0=t[cxt];
-#ifdef SM   
-    // count, prediction
-    p0+=(y<<19)-(p0>>13);
-#else
-    int pr=p0>>10;  // count, prediction
-    p0+=(((y<<22)-pr)>>3);
-#endif
-    t[cxt]=p0;
-  }
-  inline void update2(const int y, int limit) {
-    assert(cxt>=0 && cxt<N);
-    assert(y==0 || y==1);
-    U32 *p=&t[cxt], p0=p[0];
- 
-    int n=p0&1023, pr=p0>>10;  // count, prediction
-    p0+=(n<limit);
-    p0+=(((y<<22)-pr)>>3)*dt[n]&0xfffffc00;
- 
-    p[0]=p0;
-  }
-public:
-  StateMap(int n=256);
-  ~StateMap();
-  void Reset(int Rate=0){
-    for (int i=0; i<N; ++i)
-      t[i]=(t[i]&0xfffffc00)|min(Rate, t[i]&0x3FF);
-  }
-  // update bit y (0..1), predict next bit in context cx
-  int p(int cx,const int y,int limit=1023) {
-    assert(cx>=0 && cx<N);
-    assert(limit>0 && limit<1024);
-    assert(y==0 || y==1);
-    update(y,limit);
-
-    return t[cxt=cx]>>20;
-  }
-  int p1(int cx,const int y,int limit=1023) {
-    assert(cx>=0 && cx<N);
-    assert(limit>0 && limit<1024);
-    assert(y==0 || y==1);
-    update1(y,limit);
-    return t[cxt=cx]>>20;
-  }
-  int p2(int cx,const int y,int limit=1023) {
-    assert(cx>=0 && cx<N);
-    assert(limit>0 && limit<1024);
-    assert(y==0 || y==1);
-    update2(y,limit);
-    return t[cxt=cx]>>20;
-  }
-};
-StateMap::~StateMap() {
-/*    if (N==256){
-        for (U32 cx=0; cx<255; ++cx) {
-            printf ("%d ",t[cx] >> 20);
-   }
-   printf ("\n" ); }*/
-}
-
-StateMap::StateMap(int n): N(n), cxt(0), t(n) {
-  for (U32 cx=0; cx<N; ++cx) {
-        U8 state=U8(cx&255);
-        U32 n0=nex(state,2)*3+1;
-        U32 n1=nex(state,3)*3+1;
-        t[cx] = ((n1<<20)/(n0+n1))<<12;
-  }
-}
-
-// An APM maps a probability and a context to a new probability.  Methods:
-//
-// APM a(n) creates with n contexts using 96*n bytes memory.
-// a.pp(y, pr, cx, limit) updates and returns a new probability (0..4095)
-//     like with StateMap.  pr (0..4095) is considered part of the context.
-//     The output is computed by interpolating pr into 24 ranges nonlinearly
-//     with smaller ranges near the ends.  The initial output is pr.
-//     y=(0..1) is the last bit.  cx=(0..n-1) is the other context.
-//     limit=(0..1023) defaults to 255.
-
-class APM: public StateMap {
-    int steps;
-public:
-  APM(int n,const int s);
-  int p(int pr, int cx,const  int y,int limit=255) {
-    assert(pr>=0 && pr<4096);
-    assert(cx>=0 && cx<N/steps);
-    assert(y==0 || y==1);
-    assert(limit>0 && limit<1024);
-    update2(y,limit);
-    pr=(stretch(pr)+2048)*(steps-1);
-    int wt=pr&0xfff;  // interpolation weight of next element
-    cx=cx*steps+(pr>>12);
-    assert(cx>=0 && cx<N-1);
-    cxt=cx+(wt>>11);
-    pr=((t[cx]>>13)*(0x1000-wt)+(t[cx+1]>>13)*wt)>>19;
-    return pr;
-  }
-};
-
-APM::APM(int n,const int s=24): StateMap(n*s),steps(s) {
-  for (int i=0; i<N; ++i) {
-    int p=((i%steps*2+1)*4096)/(steps*2)-2048;
-    t[i]=(U32(squash(p))<<20)+6;
-  }
-}
-
 
 ///////////////////////////// BH ////////////////////////////////
 
@@ -1949,7 +1546,7 @@ public:
     lastByte = lastBit = 0;
     bits = 1;  bitPos = 0;
 #ifdef VERBOSE
-    printf("Model: %s\n",modelNames[model]);
+    printf("Model: %s\n",modelNames[model].c_str());
     printf("  CM total inputs: %d\n",inputCount);
     printf("                   MR R2 R1 R0 M1 M2 M3 M4 M5 M12 M6 E1 E2 E3 E4 E5 \n");
     printf("  CM active inputs  %d  %d  %d  %d  %d  %d  %d  %d  %d   %d  %d  %d  %d  %d  %d  %d\n",
@@ -12127,7 +11724,7 @@ Predictors():  mixerInputs(0),mixerNets(0),mixerNetsCount(0){
       // create active models
       for (int i=0;i<count;i++){
 #ifdef VERBOSE      
-          printf("Creating %s\n",modelNames[amodel[i]]);
+          printf("Creating %s\n",modelNames[amodel[i]].c_str());
 #endif
       switch (amodel[i]){
           case M_RECORD:{
